Refactor: Change package hierarchy
