+++ src/main/java/com/petspace/dev/dto/user/UserJoinRequestDto.java
@@ -1,4 +1,4 @@
-package com.petspace.dev.dto;
+package com.petspace.dev.dto.user;
 
 import com.petspace.dev.domain.OauthProvider;
 import com.petspace.dev.domain.Status;
